Float en precio articulo

+++ src/entidad/Articulo.java
@@ -32,18 +32,18 @@ public class Articulo {
 	@ManyToOne(cascade = { CascadeType.ALL })
 	@JoinColumn(name="TipoA")
 	private Tipo tipoA;
-	private int PrecioUnitario;
+	private float PrecioUnitario;
 
 	@Autowired
 	private boolean status; // 0 inactivo, 1 activo
 	
-	public int isPrecioUnitario() {
+	public float isPrecioUnitario() {
 		return PrecioUnitario;
 	}
 
 
 
-	public void setPrecioUnitario(int precioUnitario) {
+	public void setPrecioUnitario(float precioUnitario) {
 		PrecioUnitario = precioUnitario;
 	}
